+++ tests/szengine_test.py
@@ -4,7 +4,7 @@
 TODO: szengine_test.py
 """
 
-from typing import Any, Dict, List, Optional, Tuple, Union
+from typing import Any, List, Optional, Tuple
 
 import pytest
 
@@ -150,11 +150,6 @@ def test_reevaluate_record(sz_engine: SzEngineAbstract) -> None:
     sz_engine.reevaluate_record("", "")
 
 
-def test_reinitialize(sz_engine: SzEngineAbstract) -> None:
-    """Test SzEngine().reinit()."""
-    sz_engine.reinitialize(0)
-
-
 def test_search_by_attributes(sz_engine: SzEngineAbstract) -> None:
     """Test SzEngine().search_by_attributes()."""
     sz_engine.search_by_attributes("")
@@ -227,9 +222,6 @@ class SzEngineTest(SzEngineAbstract):
         **kwargs: Any,
     ) -> str:
         return ""
-
-    def destroy(self, **kwargs: Any) -> None:
-        """None"""
 
     def export_csv_entity_report(
         self,
@@ -356,16 +348,6 @@ class SzEngineTest(SzEngineAbstract):
     ) -> str:
         return ""
 
-    def initialize(
-        self,
-        instance_name: str,
-        settings: Union[str, Dict[Any, Any]],
-        config_id: Optional[int] = 0,
-        verbose_logging: Optional[int] = 0,
-        **kwargs: Any,
-    ) -> None:
-        """None"""
-
     def preprocess_record(
         self,
         record_definition: str,
@@ -391,9 +373,6 @@ class SzEngineTest(SzEngineAbstract):
         **kwargs: Any,
     ) -> str:
         return ""
-
-    def reinitialize(self, config_id: int, **kwargs: Any) -> None:
-        """None"""
 
     def search_by_attributes(
         self,