+++ tests/szerror_test.py
@@ -15,28 +15,13 @@ from senzing_abstract import SzProductAbstract
 # -----------------------------------------------------------------------------
 
 
-# def test_destroy(sz_product: SzProductAbstract) -> None:
-#     """Test SzProduct().destroy()."""
-#     sz_product.destroy()
-
-
-# def test_initialize(sz_product: SzProductAbstract) -> None:
-#     """Test SzProduct().initialize()."""
-#     sz_product.initialize("", "")
-
-
-def test_get_license(sz_product: SzProductAbstract) -> None:
+def test_new_szexception(sz_product: SzProductAbstract) -> None:
     """Test SzProduct().get_license()."""
-    sz_product.get_license()
-
-
-def test_get_version(sz_product: SzProductAbstract) -> None:
-    """Test SzProduct().get_version()."""
-    sz_product.get_version()
+    _ = sz_product
 
 
 # -----------------------------------------------------------------------------
-# SzConfig fixtures
+# SzError fixtures
 # -----------------------------------------------------------------------------