+++ tests/szhelpers_test.py
@@ -15,24 +15,14 @@ from senzing_abstract import SzProductAbstract
 # -----------------------------------------------------------------------------
 
 
-# def test_destroy(sz_product: SzProductAbstract) -> None:
-#     """Test SzProduct().destroy()."""
-#     sz_product.destroy()
-
-
-# def test_initialize(sz_product: SzProductAbstract) -> None:
-#     """Test SzProduct().initialize()."""
-#     sz_product.initialize("", "")
-
-
-def test_get_license(sz_product: SzProductAbstract) -> None:
+def test_construct_help_1(sz_product: SzProductAbstract) -> None:
     """Test SzProduct().get_license()."""
-    sz_product.get_license()
+    sz_product.help()
 
 
-def test_get_version(sz_product: SzProductAbstract) -> None:
-    """Test SzProduct().get_version()."""
-    sz_product.get_version()
+def test_construct_help_2(sz_product: SzProductAbstract) -> None:
+    """Test SzProduct().get_license()."""
+    sz_product.help("get_version")
 
 
 # -----------------------------------------------------------------------------